Move components under "shared" folder
