+++ src/routes/router.tsx
@@ -2,7 +2,7 @@ import { createBrowserRouter } from "react-router-dom";
 import TestPage from "./shared/TestPage.tsx";
 import App from "./App.tsx"
 import EventPage from "./events/EventPage.tsx";
-import ToastDemoPage from "./demo/ToastDemoPage.tsx";
+import ToastDemoPage from "./shared/ToastDemoPage.tsx";
 
 
 const router = createBrowserRouter([
